チケットmodalの修正
