+++ src/resources/sass/_variables.scss
@@ -18,6 +18,7 @@ $color: (
   lightGreen : #56b145,
   deepYellow : #FDC830,
   japanRed   : #C41A41,
+  blue       : #0179D5,
 
   twitter  : #00acee,
   facebook : #3b5998,
